Move `new_partial` & `new_partial_basics` to `partial.rs` with zero other changes
