+++ polkadot/node/service/src/lib.rs
@@ -31,6 +31,9 @@ pub mod overseer;
 pub mod workers;
 
 #[cfg(feature = "full-node")]
+mod partial;
+
+#[cfg(feature = "full-node")]
 pub use self::overseer::{
 	CollatorOverseerGen, ExtendedOverseerGenArgs, OverseerGen, OverseerGenArgs,
 	ValidatorOverseerGen,
@@ -38,6 +41,9 @@ pub use self::overseer::{
 
 #[cfg(test)]
 mod tests;
+
+#[cfg(feature = "full-node")]
+use crate::partial::{new_partial, new_partial_basics};
 
 #[cfg(feature = "full-node")]
 use {
@@ -57,7 +63,6 @@ use {
 		request_response::ReqProtocolNames,
 	},
 	sc_client_api::BlockBackend,
-	sc_consensus_grandpa::{self, FinalityProofProvider as GrandpaFinalityProofProvider},
 	sc_transaction_pool_api::OffchainTransactionPoolFactory,
 };
 
@@ -78,12 +83,9 @@ pub use {
 use std::{collections::HashMap, path::PathBuf, sync::Arc, time::Duration};
 
 use prometheus_endpoint::Registry;
-#[cfg(feature = "full-node")]
-use sc_service::KeystoreContainer;
 use sc_service::{RpcHandlers, SpawnTaskHandle};
-use sc_telemetry::TelemetryWorker;
 #[cfg(feature = "full-node")]
-use sc_telemetry::{Telemetry, TelemetryWorkerHandle};
+use sc_telemetry::TelemetryWorkerHandle;
 
 pub use chain_spec::{GenericChainSpec, RococoChainSpec, WestendChainSpec};
 use frame_benchmarking_cli::SUBSTRATE_REFERENCE_HARDWARE;
@@ -93,7 +95,7 @@ pub use polkadot_primitives::{Block, BlockId, BlockNumber, CollatorPair, Hash, I
 pub use sc_client_api::{Backend, CallExecutor};
 pub use sc_consensus::{BlockImport, LongestChain};
 pub use sc_executor::NativeExecutionDispatch;
-use sc_executor::{HeapAllocStrategy, WasmExecutor, DEFAULT_HEAP_ALLOC_STRATEGY};
+use sc_executor::WasmExecutor;
 pub use sc_service::{
 	config::{DatabaseSource, PrometheusConfig},
 	ChainSpec, Configuration, Error as SubstrateServiceError, PruningMode, Role, TFullBackend,
@@ -359,252 +361,6 @@ pub fn open_database(db_source: &DatabaseSource) -> Result<Arc<dyn Database>, Er
 		},
 	};
 	Ok(parachains_db)
-}
-
-#[cfg(feature = "full-node")]
-type FullSelectChain = relay_chain_selection::SelectRelayChain<FullBackend>;
-#[cfg(feature = "full-node")]
-type FullGrandpaBlockImport<ChainSelection = FullSelectChain> =
-	sc_consensus_grandpa::GrandpaBlockImport<FullBackend, Block, FullClient, ChainSelection>;
-#[cfg(feature = "full-node")]
-type FullBeefyBlockImport<InnerBlockImport, AuthorityId> =
-	sc_consensus_beefy::import::BeefyBlockImport<
-		Block,
-		FullBackend,
-		FullClient,
-		InnerBlockImport,
-		AuthorityId,
-	>;
-
-#[cfg(feature = "full-node")]
-struct Basics {
-	task_manager: TaskManager,
-	client: Arc<FullClient>,
-	backend: Arc<FullBackend>,
-	keystore_container: KeystoreContainer,
-	telemetry: Option<Telemetry>,
-}
-
-#[cfg(feature = "full-node")]
-fn new_partial_basics(
-	config: &mut Configuration,
-	telemetry_worker_handle: Option<TelemetryWorkerHandle>,
-) -> Result<Basics, Error> {
-	let telemetry = config
-		.telemetry_endpoints
-		.clone()
-		.filter(|x| !x.is_empty())
-		.map(move |endpoints| -> Result<_, sc_telemetry::Error> {
-			let (worker, mut worker_handle) = if let Some(worker_handle) = telemetry_worker_handle {
-				(None, worker_handle)
-			} else {
-				let worker = TelemetryWorker::new(16)?;
-				let worker_handle = worker.handle();
-				(Some(worker), worker_handle)
-			};
-			let telemetry = worker_handle.new_telemetry(endpoints);
-			Ok((worker, telemetry))
-		})
-		.transpose()?;
-
-	let heap_pages = config
-		.executor
-		.default_heap_pages
-		.map_or(DEFAULT_HEAP_ALLOC_STRATEGY, |h| HeapAllocStrategy::Static { extra_pages: h as _ });
-
-	let executor = WasmExecutor::builder()
-		.with_execution_method(config.executor.wasm_method)
-		.with_onchain_heap_alloc_strategy(heap_pages)
-		.with_offchain_heap_alloc_strategy(heap_pages)
-		.with_max_runtime_instances(config.executor.max_runtime_instances)
-		.with_runtime_cache_size(config.executor.runtime_cache_size)
-		.build();
-
-	let (client, backend, keystore_container, task_manager) =
-		sc_service::new_full_parts::<Block, RuntimeApi, _>(
-			&config,
-			telemetry.as_ref().map(|(_, telemetry)| telemetry.handle()),
-			executor,
-		)?;
-	let client = Arc::new(client);
-
-	let telemetry = telemetry.map(|(worker, telemetry)| {
-		if let Some(worker) = worker {
-			task_manager.spawn_handle().spawn(
-				"telemetry",
-				Some("telemetry"),
-				Box::pin(worker.run()),
-			);
-		}
-		telemetry
-	});
-
-	Ok(Basics { task_manager, client, backend, keystore_container, telemetry })
-}
-
-#[cfg(feature = "full-node")]
-fn new_partial<ChainSelection>(
-	config: &mut Configuration,
-	Basics { task_manager, backend, client, keystore_container, telemetry }: Basics,
-	select_chain: ChainSelection,
-) -> Result<
-	sc_service::PartialComponents<
-		FullClient,
-		FullBackend,
-		ChainSelection,
-		sc_consensus::DefaultImportQueue<Block>,
-		sc_transaction_pool::TransactionPoolHandle<Block, FullClient>,
-		(
-			impl Fn(
-				polkadot_rpc::SubscriptionTaskExecutor,
-			) -> Result<polkadot_rpc::RpcExtension, SubstrateServiceError>,
-			(
-				sc_consensus_babe::BabeBlockImport<
-					Block,
-					FullClient,
-					FullBeefyBlockImport<
-						FullGrandpaBlockImport<ChainSelection>,
-						ecdsa_crypto::AuthorityId,
-					>,
-				>,
-				sc_consensus_grandpa::LinkHalf<Block, FullClient, ChainSelection>,
-				sc_consensus_babe::BabeLink<Block>,
-				sc_consensus_beefy::BeefyVoterLinks<Block, ecdsa_crypto::AuthorityId>,
-			),
-			sc_consensus_grandpa::SharedVoterState,
-			sp_consensus_babe::SlotDuration,
-			Option<Telemetry>,
-		),
-	>,
-	Error,
->
-where
-	ChainSelection: 'static + SelectChain<Block>,
-{
-	let transaction_pool = Arc::from(
-		sc_transaction_pool::Builder::new(
-			task_manager.spawn_essential_handle(),
-			client.clone(),
-			config.role.is_authority().into(),
-		)
-		.with_options(config.transaction_pool.clone())
-		.with_prometheus(config.prometheus_registry())
-		.build(),
-	);
-
-	let grandpa_hard_forks = if config.chain_spec.is_kusama() {
-		grandpa_support::kusama_hard_forks()
-	} else {
-		Vec::new()
-	};
-
-	let (grandpa_block_import, grandpa_link) =
-		sc_consensus_grandpa::block_import_with_authority_set_hard_forks(
-			client.clone(),
-			GRANDPA_JUSTIFICATION_PERIOD,
-			&(client.clone() as Arc<_>),
-			select_chain.clone(),
-			grandpa_hard_forks,
-			telemetry.as_ref().map(|x| x.handle()),
-		)?;
-	let justification_import = grandpa_block_import.clone();
-
-	let (beefy_block_import, beefy_voter_links, beefy_rpc_links) =
-		sc_consensus_beefy::beefy_block_import_and_links(
-			grandpa_block_import,
-			backend.clone(),
-			client.clone(),
-			config.prometheus_registry().cloned(),
-		);
-
-	let babe_config = sc_consensus_babe::configuration(&*client)?;
-	let (block_import, babe_link) =
-		sc_consensus_babe::block_import(babe_config.clone(), beefy_block_import, client.clone())?;
-
-	let slot_duration = babe_link.config().slot_duration();
-	let (import_queue, babe_worker_handle) =
-		sc_consensus_babe::import_queue(sc_consensus_babe::ImportQueueParams {
-			link: babe_link.clone(),
-			block_import: block_import.clone(),
-			justification_import: Some(Box::new(justification_import)),
-			client: client.clone(),
-			select_chain: select_chain.clone(),
-			create_inherent_data_providers: move |_, ()| async move {
-				let timestamp = sp_timestamp::InherentDataProvider::from_system_time();
-
-				let slot =
-				sp_consensus_babe::inherents::InherentDataProvider::from_timestamp_and_slot_duration(
-					*timestamp,
-					slot_duration,
-				);
-
-				Ok((slot, timestamp))
-			},
-			spawner: &task_manager.spawn_essential_handle(),
-			registry: config.prometheus_registry(),
-			telemetry: telemetry.as_ref().map(|x| x.handle()),
-			offchain_tx_pool_factory: OffchainTransactionPoolFactory::new(transaction_pool.clone()),
-		})?;
-
-	let justification_stream = grandpa_link.justification_stream();
-	let shared_authority_set = grandpa_link.shared_authority_set().clone();
-	let shared_voter_state = sc_consensus_grandpa::SharedVoterState::empty();
-	let finality_proof_provider = GrandpaFinalityProofProvider::new_for_service(
-		backend.clone(),
-		Some(shared_authority_set.clone()),
-	);
-
-	let import_setup = (block_import, grandpa_link, babe_link, beefy_voter_links);
-	let rpc_setup = shared_voter_state.clone();
-
-	let rpc_extensions_builder = {
-		let client = client.clone();
-		let keystore = keystore_container.keystore();
-		let transaction_pool = transaction_pool.clone();
-		let select_chain = select_chain.clone();
-		let chain_spec = config.chain_spec.cloned_box();
-		let backend = backend.clone();
-
-		move |subscription_executor: polkadot_rpc::SubscriptionTaskExecutor|
-		      -> Result<polkadot_rpc::RpcExtension, sc_service::Error> {
-			let deps = polkadot_rpc::FullDeps {
-				client: client.clone(),
-				pool: transaction_pool.clone(),
-				select_chain: select_chain.clone(),
-				chain_spec: chain_spec.cloned_box(),
-				babe: polkadot_rpc::BabeDeps {
-					babe_worker_handle: babe_worker_handle.clone(),
-					keystore: keystore.clone(),
-				},
-				grandpa: polkadot_rpc::GrandpaDeps {
-					shared_voter_state: shared_voter_state.clone(),
-					shared_authority_set: shared_authority_set.clone(),
-					justification_stream: justification_stream.clone(),
-					subscription_executor: subscription_executor.clone(),
-					finality_provider: finality_proof_provider.clone(),
-				},
-				beefy: polkadot_rpc::BeefyDeps::<ecdsa_crypto::AuthorityId> {
-					beefy_finality_proof_stream: beefy_rpc_links.from_voter_justif_stream.clone(),
-					beefy_best_block_stream: beefy_rpc_links.from_voter_best_beefy_stream.clone(),
-					subscription_executor,
-				},
-				backend: backend.clone(),
-			};
-
-			polkadot_rpc::create_full(deps).map_err(Into::into)
-		}
-	};
-
-	Ok(sc_service::PartialComponents {
-		client,
-		backend,
-		task_manager,
-		keystore_container,
-		select_chain,
-		import_queue,
-		transaction_pool,
-		other: (rpc_extensions_builder, import_setup, rpc_setup, slot_duration, telemetry),
-	})
 }
 
 #[cfg(feature = "full-node")]
